# Fix of unexpected MatchError in case of invalid escape character