--- a/version.sbt
+++ b/version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "1.29.1"
+version in ThisBuild := "1.29.2-SNAPSHOT"