@plokhotnyuk 2018-09-09
Should I add a `travis_wait 30` before `sbt ...` command in the `.travis.yml` file to overcome 10 min of build timeout on Travis CI?