@ghik 2018-09-11
Thanks a lot for the improvements. I normally don't use unsafe matches except for "raw" code like this one where performance is a priority. I don't know why the build hanged up - you can ignore it.